1. Синхронный обмен сообщений [#458496 #307100]

+++ src/main/java/ru/job4j/service/PassportService.java
@@ -1,11 +1,14 @@
 package ru.job4j.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import ru.job4j.model.Passport;
 import ru.job4j.repository.PassportRepository;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class PassportService {
@@ -14,16 +17,22 @@ public class PassportService {
     private PassportRepository passportRepository;
 
      public Passport save(Passport passport) {
-
+         if (!findPassportsBySeries(passport.getSeries()).isEmpty()
+                 || findPassportsByNumber(passport.getNumber()).isEmpty()) {
+             throw new IllegalArgumentException();
+         }
       return passportRepository.save(passport);
      }
 
      public Passport update(int id, Passport passport) {
          Passport passportUpdate = passportRepository.findPassportById(id);
+         if (passportUpdate == null) {
+             throw new NoSuchElementException();
+         }
          passportUpdate.setId(id);
          passportUpdate.setName(passport.getName());
          passportUpdate.setLastname(passport.getLastname());
-         passportUpdate.setCreated(passport.getCreated());
+         passportUpdate.setExpirationDate(passport.getExpirationDate());
          return passportRepository.save(passportUpdate);
      }
 
@@ -32,6 +41,9 @@ public class PassportService {
      }
 
     public void delete(int id) {
+        if (findPassportById(id) == null) {
+            throw new NoSuchElementException();
+        }
          Passport passport = findPassportById(id);
          passportRepository.delete(passport);
     }
